Improved map userability
only bound the map on the latitude. Allow "infinite" scrolling on x
axis.

Leaflet does not move the markers as you pan. The markers are drawn
+-180 degrees lng. When we pan, we need to remove the markers from the
oppsite site and place them in the direction we are panning. This gives
the experience of a "globe".

NOTE: Performance is sufficient on a 2012 Macbook air with ~2000
markers. We may need to revisit this when more markers are rendered.